Factor out Nepali date functionality
